api endpoints on user signup
