--- api/models/user.js
@@ -1,13 +1,0 @@
-const users = [
-  {
-    id: '001',
-    firstName: 'Celestin',
-    lastName: 'NIYONSABA',
-    email: 'niyoceles3@gmail.com',
-    phone: '+250783067644',
-    userName: 'niyoceles',
-    password: 'celes123',
-  },
-];
-
-export default users;
