+++ api/models/users.js
@@ -1,0 +1,31 @@
+const users = [
+  {
+    id: 1,
+    firstName: 'Celestin',
+    lastName: 'NIYONSABA',
+    email: 'niyoceles3@gmail.com',
+    phone: '+250783067644',
+    userName: 'niyoceles',
+    password: 'celes123',
+  },
+  {
+    id: 2,
+    firstName: 'John',
+    lastName: 'MAFI',
+    email: 'mafi@gmail.com',
+    phone: '+2507888888',
+    userName: 'mafi',
+    password: '12345',
+  },
+  {
+    id: 3,
+    firstName: 'Alex',
+    lastName: 'KALISA',
+    email: 'alexk@gmail.com',
+    phone: '+2507899999',
+    userName: 'kalisa',
+    password: 'alex123',
+  },
+];
+
+export default users;
