+++ api/routes/users.js
@@ -1,9 +1,10 @@
 import { Router } from 'express';
-import User from '../controllers/User';
+import User from '../controllers/UsersController';
 
 const router = Router();
 
-router.get('/users', User.getUsers);
+router.get('/users', User.getAllusers);
+router.get('/users/:id', User.getSingleUser);
 router.post('/signup', User.signup);
 router.post('/signin', User.signin);
